# docs/phase1/04-openapi-specification.yaml
openapi: 3.0.3
info:
  title: CardDemo REST API
  description: |
    REST API specification for the CardDemo credit card management application,
    migrated from COBOL/CICS mainframe to Java 17 Spring Boot.
    
    This API replaces the following mainframe technologies:
    - CICS online transactions → REST endpoints
    - BMS screens → JSON request/response
    - VSAM file I/O → JPA/PostgreSQL
    - RACF security → JWT authentication
  version: 1.0.0
  contact:
    name: CardDemo Migration Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticates a user and returns a JWT token.
        Replaces CICS transaction CC00 (COSGN00C).
        Validates against USER_SECURITY table.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: Invalidates the current JWT token.
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
...
  /api/transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      description: |
        Returns paginated list of transactions with filtering.
        Replaces CICS transaction CT00 (COTRN00C).
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: cardNumber
          in: query
          schema:
            type: string
          description: Filter by card number
        - name: accountId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by account ID
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
    post:
      tags: [Transactions]
      summary: Add transaction
      description: |
        Creates a new transaction.
        Replaces CICS transaction CT02 (COTRN02C).
      operationId: addTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Validation error
  /api/transactions/{transactionId}:
    get:
      tags: [Transactions]
      summary: View transaction
      description: |
        Retrieves transaction details.
        Replaces CICS transaction CT01 (COTRN01C).
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
...
  /api/reports/transactions:
    post:
      tags: [Reports]
      summary: Generate transaction report
      description: |
        Generates a transaction report for a date range.
        Replaces CICS transaction CR00 (CORPT00C) and batch job TRANREPT (CBTRN03C).
      operationId: generateTransactionReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionReportRequest'
      responses:
        '200':
          description: Report generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionReportResponse'
...
  /api/admin/transaction-types:
    get:
      tags: [Admin - Transaction Types]
      summary: List transaction types
      description: |
        Returns paginated list of transaction types.
        Replaces CICS transaction CTLI (COTRTLIC).
      operationId: listTransactionTypes
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Transaction type list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypeListResponse'
    post:
      tags: [Admin - Transaction Types]
      summary: Add transaction type
      description: |
        Creates a new transaction type.
        Replaces CICS transaction CTTU (COTRTUPC) add function.
      operationId: addTransactionType
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeRequest'
      responses:
        '201':
          description: Transaction type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypeResponse'
  /api/admin/transaction-types/{typeCode}:
    put:
      tags: [Admin - Transaction Types]
      summary: Update transaction type
      description: Replaces CICS transaction CTTU (COTRTUPC) edit function.
      operationId: updateTransactionType
      security:
        - bearerAuth: []
      parameters:
        - name: typeCode
          in: path
          required: true
          schema:
            type: string
            maxLength: 2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeRequest'
      responses:
        '200':
          description: Transaction type updated
    delete:
      tags: [Admin - Transaction Types]
      summary: Delete transaction type
      description: Replaces CICS transaction CTLI (COTRTLIC) delete function.
      operationId: deleteTransactionType
      security:
        - bearerAuth: []
      parameters:
        - name: typeCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Transaction type deleted
        '409':
          description: Cannot delete - referenced by transaction categories
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [userId, password]
      properties:
        userId:
          type: string
          maxLength: 8
          example: USER0001
        password:
          type: string
          maxLength: 8
          example: PASSWORD
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        userId:
          type: string
        userType:
          type: string
          enum: [ADMIN, USER]
        firstName:
          type: string
        lastName:
          type: string
...
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        currentPage:
          type: integer
    TransactionResponse:
      type: object
      properties:
        transactionId:
          type: string
        typeCode:
          type: string
        categoryCode:
          type: integer
        source:
          type: string
        description:
          type: string
        amount:
          type: number
          format: double
        merchantId:
          type: integer
          format: int64
        merchantName:
          type: string
        merchantCity:
          type: string
        merchantZip:
          type: string
        cardNumber:
          type: string
        origTimestamp:
          type: string
          format: date-time
        procTimestamp:
          type: string
          format: date-time
    TransactionCreateRequest:
      type: object
      required: [typeCode, categoryCode, cardNumber, amount]
      properties:
        typeCode:
          type: string
          maxLength: 2
        categoryCode:
          type: integer
        source:
          type: string
          maxLength: 10
        description:
          type: string
          maxLength: 100
        amount:
          type: number
          format: double
        merchantId:
          type: integer
          format: int64
        merchantName:
          type: string
          maxLength: 50
        merchantCity:
          type: string
          maxLength: 50
        merchantZip:
          type: string
          maxLength: 10
        cardNumber:
          type: string
          maxLength: 16
...
    TransactionReportRequest:
      type: object
      required: [startDate, endDate]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        accountId:
          type: integer
          format: int64
          description: Optional filter by account
    TransactionReportResponse:
      type: object
      properties:
        reportName:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        pageTotal:
          type: number
          format: double
        grandTotal:
          type: number
          format: double
...
    TransactionTypeListResponse:
      type: object
      properties:
        transactionTypes:
          type: array
          items:
            $ref: '#/components/schemas/TransactionTypeResponse'
        totalElements:
          type: integer
          format: int64
    TransactionTypeResponse:
      type: object
      properties:
        typeCode:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: object
            properties:
              categoryCode:
                type: string
              description:
                type: string
    TransactionTypeRequest:
      type: object
      required: [typeCode, description]
      properties:
        typeCode:
          type: string
          maxLength: 2
        description:
          type: string
          maxLength: 50
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
security:
  - bearerAuth: []